requirements.txt
pyautogen
python-dotenv
pydantic